chore(dead-code): archive the useRecommendation subtree; split the live FragileEdgeGroupCard out of the dead ChallengeSection
Two dead-code removals, both re-derived at this tip rather than inherited
from their manifests.

PART 1 — archive the useRecommendation / RecommendationCard subtree (22 files)

Zero production importers. Verdict and served-bundle evidence in
docs-designs/USERECOMMENDATION-REACHABILITY-2026-07-27.md (pinned 13cca490);
the cascade was RE-DERIVED here at 6d474415 from an exact reverse-dependency
map (every import specifier resolved to a real path, then inverted), because
four name collisions make a symbol grep wrong on this subtree — two
DriversSection.tsx, two ExpandableSection.tsx, useRobustness (canvas hook vs
an unrelated inspector-v2 export), and RobustnessIndicator, whose JSX mount at
InlineBlocks.tsx:752 resolves to a LOCAL function declared at :815, not to
this component.

The cascade runs WIDER than the pinned doc's co-fate list: it also takes
components/ConditionalGuidance/ (both files), hooks/useConditionalRecommendations.ts,
and utils/{coherenceCheck,cleanInsightText}.ts, each of which retained only
its own spec once index.tsx died.

KEPT, deliberately: RecommendationCard/types.ts (type-only, zero runtime
exports, but 10 live importers — moving it breaks the build; this is the one
trap in the slice) and hooks/useRobustness.ts (still used by OutputsDock and
useResultsPanelData).

Stated consequence: useRecommendation.ts held the UI's ONLY call site of
/bff/engine/v1/recommend/generate. That endpoint is now UI-unconsumed. This
is deliberate — do not read the UI's silence as evidence no service calls it.

PART 2 — split ChallengeSection: the component is dead, the module is not

ChallengeSection.tsx exported a dead component AND a live one. The live chain
died at ResultsBody.tsx when StressTestSection replaced the Challenge
accordion, but FragileEdgeGroupCard is mounted by StressTestSection.tsx:284 on
the live V5 path and ChallengeFragileEdge is a live wire type.

FragileEdgeGroupCard + ChallengeFragileEdge move to their own module,
BYTE-IDENTICAL (verified: the 15-line interface and 127-line function diff
clean against the original) — #508's fragile-edge copy, its shape pins, and
the REQUIRED designationsWithheld prop threading are untouched. Deleted: the
ChallengeSection component, ChallengeCard (with its ungated "Run this exercise
to challenge the recommendation." string), IdentifiabilityCard, and CappedList
(whose only production importer was ChallengeSection) with its spec.

Test coverage was PORTED, not dropped — 13 of 13 preserved:
  * 12 of the 13 fragileRows mounts re-point at FragileEdgeGroupCard directly.
  * The 13th — the "(Status Quo)" suffix strip — is CALLER behaviour
    (stripStatusQuoSuffixForDisplay runs in the grouping reduce, never in the
    card), so it moved to StressTestSection.spec.tsx on the live path.
    Re-pointing it at the card would have made it vacuous: the test would have
    had to pre-strip the label and then assert the strip happened.

CORRECTION to the Part 2 brief: the 4 mounts in v12.2-factor-labels.spec.tsx
do NOT exercise FragileEdgeGroupCard — none of them passes fragileEdges at
all. Their subject is ChallengeSection's own resolveFactorLabel + ChallengeCard,
a third independent id-prettifying fallback. They are DELETED WITH THEIR
SUBJECT, per the brief's own rule, not ported.

Two stale mirrors fixed in passing:
  * no-message-render.spec.ts exempted 'ChallengeSection.tsx' on an attested
    pattern with ZERO occurrences in the file (inference warnings had moved
    out in Brief 4 Task 12). It never fired only because the exemption is
    consulted lazily. DEFENCE_IN_DEPTH_FILES has no stale-entry check.
  * cee-rendering-claims.contract.test.tsx cited "ChallengeSection.tsx:97-99
    carries a THIRD independent id-prettifying fallback" — now two, not three.

BASELINES — all shrink except one, which is mechanical and explained
  * typecheck ratchet 2545 -> 2517. Attributed file-by-file: 2+2+1+7+3+4+4
    from the seven removed erroring files, +5 from v12.2 (10 -> 5, trimmed not
    deleted) = exactly 28. No new diagnostics in any new or edited file.
  * DS compliance: legacy-tailwind-token 2023 -> 1790, production-hex
    295 -> 289, uppercase-text 64 -> 62 (measured against a pristine-staging
    control worktree). panel-typography-scoped is UP vs its baseline (45 v 34)
    but is IDENTICALLY up on pristine 6d474415 — pre-existing, not this PR,
    and the DS baseline is deliberately NOT regenerated here.
  * scripts/ci/typecheck-uncovered.txt grows by 22 (18 -> 40 entries). This is
    the ONE count that rises. It is not a suppressed regression: the list is a
    bidirectional per-file exception list for tracked TypeScript no tsconfig
    loads, so archiving REQUIRES an entry per moved file or the gate fails.
