+++ archive/dead-recommendation-subtree-2026-07/canvas/components/RecommendationCard/ParetoMiniChart.tsx
