+++ archive/dead-recommendation-subtree-2026-07/canvas/hooks/useConditionalRecommendations.ts
