+++ archive/dead-recommendation-subtree-2026-07/canvas/utils/coherenceCheck.ts
