+++ scripts/ci/typecheck-baseline.txt
@@ -10,7 +10,7 @@
 # this file and phase 2.
 #
 # Format: <error-count><TAB><path>, sorted by path.
-# count=2545
+# count=2517
 1	e2e/_helpers.ts
 1	e2e/a11y.spec.ts
 9	e2e/canvas-inspector.spec.ts
@@ -124,7 +124,6 @@
 8	src/canvas/components/CommandPalette.tsx
 7	src/canvas/components/CompareView.tsx
 2	src/canvas/components/ComparisonTable.tsx
-2	src/canvas/components/ConditionalGuidance/index.tsx
 1	src/canvas/components/ConformalPrediction.tsx
 1	src/canvas/components/DecisionQuality.tsx
 1	src/canvas/components/DecisionReadinessBadge.tsx
@@ -152,9 +151,6 @@
 1	src/canvas/components/ParetoInsights.tsx
 27	src/canvas/components/PreAnalysisGuidance.tsx
 2	src/canvas/components/RangeDisplay.tsx
-2	src/canvas/components/RecommendationCard/RobustnessBlock.tsx
-1	src/canvas/components/RecommendationCard/RobustnessIndicator.tsx
-7	src/canvas/components/RecommendationCard/index.tsx
 3	src/canvas/components/RecoveryBanner.tsx
 2	src/canvas/components/RiskAdjustedDisplay.tsx
 2	src/canvas/components/RiskProfileBadge.tsx
@@ -264,14 +260,12 @@
 2	src/canvas/hooks/__tests__/isAnalysisReadyStale.spec.ts
 2	src/canvas/hooks/__tests__/useGuidancePulseHighlight.spec.ts
 13	src/canvas/hooks/__tests__/usePathHighlight.spec.ts
-3	src/canvas/hooks/__tests__/useRecommendation.plotBearer.spec.tsx
 2	src/canvas/hooks/__tests__/useV2Run.concurrency.spec.ts
 1	src/canvas/hooks/__tests__/useV2Run.interventionValidation.spec.ts
 1	src/canvas/hooks/useAddBaseline.ts
 7	src/canvas/hooks/useAnalysisMetadata.ts
 1	src/canvas/hooks/useAutosave.ts
 2	src/canvas/hooks/useBlueprintInsert.ts
-4	src/canvas/hooks/useConditionalRecommendations.ts
 1	src/canvas/hooks/useHighlightDispatch.ts
 1	src/canvas/hooks/useKeyInsight.ts
 4	src/canvas/hooks/useLensFilter.ts
@@ -282,7 +276,6 @@
 2	src/canvas/hooks/usePreAnalysisData.ts
 6	src/canvas/hooks/usePreRunValidation.ts
 6	src/canvas/hooks/usePreviewRun.ts
-4	src/canvas/hooks/useRecommendation.ts
 3	src/canvas/hooks/useResultsRun.ts
 8	src/canvas/hooks/useRiskProfile.ts
 3	src/canvas/hooks/useRiskToleranceSuggestion.ts
@@ -472,7 +465,7 @@
 2	src/components/results/__tests__/driverPolicyFeed.parity.spec.tsx
 1	src/components/results/__tests__/golden-payload.spec.ts
 5	src/components/results/__tests__/isl-field-surfacing.spec.tsx
-10	src/components/results/__tests__/v12.2-factor-labels.spec.tsx
+5	src/components/results/__tests__/v12.2-factor-labels.spec.tsx
 13	src/components/results/__tests__/v12.2-fixes.spec.tsx
 1	src/components/results/analysisHeroV17/__tests__/accessibility.spec.tsx
 3	src/components/results/analysisHeroV17/__tests__/bodyLabelSafety.spec.tsx
