+++ scripts/ci/typecheck-uncovered.txt
@@ -31,6 +31,28 @@ archive/dead-canvas-components-2026-07/__tests__/EmptyStateOverlay.test.tsx
 archive/dead-canvas-components-2026-07/__tests__/InputsDock.dom.spec.tsx
 archive/dead-canvas-components-2026-07/__tests__/OutcomesSignal.spec.tsx
 archive/dead-canvas-theme-2026-07/nodes.ts
+archive/dead-recommendation-subtree-2026-07/canvas/components/ConditionalGuidance/index.tsx
+archive/dead-recommendation-subtree-2026-07/canvas/components/ConditionalGuidance/types.ts
+archive/dead-recommendation-subtree-2026-07/canvas/components/RecommendationCard/AssumptionsSection.tsx
+archive/dead-recommendation-subtree-2026-07/canvas/components/RecommendationCard/ConstraintViolationsSection.tsx
+archive/dead-recommendation-subtree-2026-07/canvas/components/RecommendationCard/DriversSection.tsx
+archive/dead-recommendation-subtree-2026-07/canvas/components/RecommendationCard/ExpandableSection.tsx
+archive/dead-recommendation-subtree-2026-07/canvas/components/RecommendationCard/ParetoMiniChart.tsx
+archive/dead-recommendation-subtree-2026-07/canvas/components/RecommendationCard/RobustnessBlock.tsx
+archive/dead-recommendation-subtree-2026-07/canvas/components/RecommendationCard/RobustnessIndicator.tsx
+archive/dead-recommendation-subtree-2026-07/canvas/components/RecommendationCard/TradeoffsSection.tsx
+archive/dead-recommendation-subtree-2026-07/canvas/components/RecommendationCard/__tests__/ParetoMiniChart.spec.tsx
+archive/dead-recommendation-subtree-2026-07/canvas/components/RecommendationCard/__tests__/RecommendationCard.spec.tsx
+archive/dead-recommendation-subtree-2026-07/canvas/components/RecommendationCard/__tests__/RobustnessBlock.spec.tsx
+archive/dead-recommendation-subtree-2026-07/canvas/components/RecommendationCard/__tests__/RobustnessIndicator.spec.tsx
+archive/dead-recommendation-subtree-2026-07/canvas/components/RecommendationCard/index.tsx
+archive/dead-recommendation-subtree-2026-07/canvas/hooks/__tests__/useRecommendation.plotBearer.spec.tsx
+archive/dead-recommendation-subtree-2026-07/canvas/hooks/useConditionalRecommendations.ts
+archive/dead-recommendation-subtree-2026-07/canvas/hooks/useRecommendation.ts
+archive/dead-recommendation-subtree-2026-07/canvas/utils/__tests__/cleanInsightText.spec.ts
+archive/dead-recommendation-subtree-2026-07/canvas/utils/__tests__/coherenceCheck.spec.ts
+archive/dead-recommendation-subtree-2026-07/canvas/utils/cleanInsightText.ts
+archive/dead-recommendation-subtree-2026-07/canvas/utils/coherenceCheck.ts
 archive/dead-ui-components-2026-07/KPIHeadline.tsx
 archive/dead-ui-components-2026-07/StabilityGauge.tsx
 archive/dead-ui-components-2026-07/__tests__/StabilityGauge.test.tsx
